added update-support to prepare_boot.sh

+++ gentoo-buildchain/prepare_bootfiles.sh
@@ -12,13 +12,15 @@ https://github.com/raspberrypi/firmware/raw/master/boot/bootcode.bin
 wget -qq -O ./mnt/gentoo/boot/fixup.dat \
 https://github.com/raspberrypi/firmware/raw/master/boot/fixup.dat
 
-echo -e "patching boot/cmdline.txt..."
-echo -e "root=/dev/mmcblk0p2 rootfstype=ext4 rootdelay=2" > ./mnt/gentoo/boot/cmdline.txt
-
-echo -e "patching boot/config.txt..."
-echo -e \
-'framebuffer_depth=24
-boot_delay=1' > ./mnt/gentoo/boot/config.txt
-
+if [[ ! -f ./mnt/gentoo/boot/cmdline.txt ]]; then 
+	echo -e "patching boot/cmdline.txt..."
+	echo -e "root=/dev/mmcblk0p2 rootfstype=ext4 rootdelay=2" > ./mnt/gentoo/boot/cmdline.txt
+fi
+if [[ ! -f ./mnt/gentoo/boot/config.txt ]]; then
+	echo -e "patching boot/config.txt..."
+	echo -e \
+	'framebuffer_depth=24
+	boot_delay=1' > ./mnt/gentoo/boot/config.txt
+fi
 
 echo -e "done, have a nice day! :)"
